renumber files to ED 1-10
